pc_launcher/requirements.txt
pydantic>=2.11,<3
PyYAML>=6.0,<7
python-dotenv>=1.1,<2
requests>=2.32,<3
